third_party/blink/web_tests/platform/generic/fast/css/invalidation/targeted-class-custom-pseudo-expected.txt

Check that targeted class invalidation works for custom pseudo elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS !!window.internals is true
PASS getComputedStyle(fileUploadButton, null).backgroundColor is gray
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(fileUploadButton, null).backgroundColor is green
PASS getComputedStyle(t2, null).backgroundColor is lime
PASS successfullyParsed is true

TEST COMPLETE
